+++ packages/client-auth/src/lib/config.ts
@@ -61,7 +61,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.27",
-    genVersion: "2.396.0",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.27 2.396.0 0.0.0 @formance/sdk-auth",
+    sdkVersion: "1.0.0-alpha.28",
+    genVersion: "2.399.0",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.28 2.399.0 0.0.0 @formance/sdk-auth",
 } as const;