+++ packages/client-auth/RELEASES.md
@@ -279,3 +279,13 @@ Based on:
 - [typescript v1.0.0-alpha.27] packages/client-auth
 ### Releases
 - [NPM v1.0.0-alpha.27] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.27 - packages/client-auth
+
+## 2024-08-15 00:09:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.368.0 (2.399.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.28] packages/client-auth
+### Releases
+- [NPM v1.0.0-alpha.28] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.28 - packages/client-auth